+++ SidecarFixup/kern_start.cpp
@@ -30,38 +30,24 @@ static const uint8_t kIPadModelPatched[] = { 0x68, 0x50, 0x61, 0x64, 0x34, 0x2C,
 static_assert(sizeof(kMacModelOriginal) == sizeof(kMacModelPatched), "patch size invalid");
 static_assert(sizeof(kIPadModelOriginal) == sizeof(kIPadModelPatched), "patch size invalid");
 
-static const char *dyldSharedPath;
-
 static mach_vm_address_t orig_cs_validate {};
 
 #pragma mark - Kernel patching code
 
-template <size_t patchSize>
-static inline void searchAndPatch(const void *haystack,
-                                  size_t haystackSize,
-                                  const char *path,
-                                  const uint8_t (&needle)[patchSize],
-                                  const uint8_t (&patch)[patchSize]) {
-    if (UNLIKELY(strcmp(path, dyldSharedPath) == 0)) {
-        if (UNLIKELY(KernelPatcher::findAndReplace(const_cast<void *>(haystack), haystackSize, needle, patchSize, patch, patchSize))) {
-            DBGLOG(MODULE_SHORT, "found function to patch at %s!", path);
-        }
-    }
+template <size_t findSize, size_t replaceSize>
+static inline void searchAndPatch(const void *haystack, size_t haystackSize, const char *path, const uint8_t (&needle)[findSize], const uint8_t (&patch)[replaceSize]) {
+   if (UNLIKELY(KernelPatcher::findAndReplace(const_cast<void *>(haystack), haystackSize, needle, findSize, patch, replaceSize)))
+	   DBGLOG(MODULE_SHORT, "found function to patch at %s!", path);
 }
 
 #pragma mark - Patched functions
 
 // pre Big Sur
-static boolean_t patched_cs_validate_range(vnode_t vp,
-                                           memory_object_t pager,
-                                           memory_object_offset_t offset,
-                                           const void *data,
-                                           vm_size_t size,
-                                           unsigned *result) {
+static boolean_t patched_cs_validate_range(vnode_t vp, memory_object_t pager, memory_object_offset_t offset, const void *data, vm_size_t size, unsigned *result) {
     char path[PATH_MAX];
     int pathlen = PATH_MAX;
     boolean_t res = FunctionCast(patched_cs_validate_range, orig_cs_validate)(vp, pager, offset, data, size, result);
-    if (res && vn_getpath(vp, path, &pathlen) == 0) {
+    if (res && vn_getpath(vp, path, &pathlen) == 0 && UserPatcher::matchSharedCachePath(path)) {
         searchAndPatch(data, size, path, kMacModelOriginal, kMacModelPatched);
         searchAndPatch(data, size, path, kIPadModelOriginal, kIPadModelPatched);
     }
@@ -69,17 +55,11 @@ static boolean_t patched_cs_validate_range(vnode_t vp,
 }
 
 // For Big Sur
-static void patched_cs_validate_page(vnode_t vp,
-                                          memory_object_t pager,
-                                          memory_object_offset_t page_offset,
-                                          const void *data,
-                                          int *arg4,
-                                          int *arg5,
-                                          int *arg6) {
+static void patched_cs_validate_page(vnode_t vp, memory_object_t pager, memory_object_offset_t page_offset, const void *data, int *validated_p, int *tainted_p, int *nx_p) {
     char path[PATH_MAX];
     int pathlen = PATH_MAX;
-    FunctionCast(patched_cs_validate_page, orig_cs_validate)(vp, pager, page_offset, data, arg4, arg5, arg6);
-    if (vn_getpath(vp, path, &pathlen) == 0) {
+    FunctionCast(patched_cs_validate_page, orig_cs_validate)(vp, pager, page_offset, data, validated_p, tainted_p, nx_p);
+	if (vn_getpath(vp, path, &pathlen) == 0 && UserPatcher::matchSharedCachePath(path)) {
         searchAndPatch(data, PAGE_SIZE, path, kMacModelOriginal, kMacModelPatched);
         searchAndPatch(data, PAGE_SIZE, path, kIPadModelOriginal, kIPadModelPatched);
     }
@@ -88,16 +68,14 @@ static void patched_cs_validate_page(vnode_t vp,
 #pragma mark - Patches on start/stop
 
 static void pluginStart() {
-	dyldSharedPath = UserPatcher::getSharedCachePath();
-    DBGLOG(MODULE_SHORT, "start");
+	DBGLOG(MODULE_SHORT, "start");
 	lilu.onPatcherLoadForce([](void *user, KernelPatcher &patcher) {
 		KernelPatcher::RouteRequest csRoute =
 			getKernelVersion() >= KernelVersion::BigSur ?
 			KernelPatcher::RouteRequest("_cs_validate_page", patched_cs_validate_page, orig_cs_validate) :
 			KernelPatcher::RouteRequest("_cs_validate_range", patched_cs_validate_range, orig_cs_validate);
-		if (!patcher.routeMultipleLong(KernelPatcher::KernelID, &csRoute, 1)) {
+		if (!patcher.routeMultipleLong(KernelPatcher::KernelID, &csRoute, 1))
 			SYSLOG(MODULE_SHORT, "failed to route cs validation pages");
-		}
 	});
 }